fix docker-compose, dockerfile, adding local feature toggling service

+++ src/main/kotlin/com/example/minibank/controllers/AuthController.kt
@@ -3,6 +3,7 @@ package com.example.minibank.controllers
 import com.example.minibank.dtos.UserDTO
 import com.example.minibank.dtos.exeptions.Message
 import com.example.minibank.models.User
+import com.example.minibank.services.FTService
 import com.example.minibank.services.UserService
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
@@ -25,17 +26,39 @@ import javax.servlet.http.HttpServletResponse
     description = "Все методы для работы с пользователями",
 )
 
-class AuthController(private val userService: UserService) {
+class AuthController(
+    private val userService: UserService,
+    val ftService: FTService
+) {
+
+    private val userCanCreate: String = "userCanCreate"
+    private val userCanGet: String = "userCanGet"
+    private val userCanDelete: String = "userCanDelete"
+    private val userCanLogin: String = "userCanLogin"
+    private val userCanUpdate: String = "userCanUpdate"
+    private val userCanlogout: String = "logout"
 
     @PutMapping("/user/{id}")
     @Operation(summary = "Обновление данных пользователя")
-    fun updateUser(@PathVariable id: Int, @RequestBody user: UserDTO) {
+    fun updateUser(@PathVariable id: Int, @RequestBody user: UserDTO): Any {
+        if (!ftService.isEnabled(userCanUpdate)) {
+            println("feature $userCanUpdate is off")
+            return ResponseEntity.ok().body(Message("FT $userCanUpdate is off"))
+        }
+        println("feature $userCanUpdate is on")
+
         return userService.updateUser(id, user)
     }
 
     @PostMapping("/user")
     @Operation(summary = "Создание пользователя")
     fun register(@RequestBody body: UserDTO): ResponseEntity<Any> {
+        if (!ftService.isEnabled(userCanCreate)) {
+            println("feature $userCanCreate is off")
+            return ResponseEntity.ok().body(Message("FT $userCanCreate is off"))
+        }
+        println("feature $userCanCreate is on")
+
         return try {
             val user = User()
             @Operation(description = "Имя ")
@@ -56,17 +79,32 @@ class AuthController(private val userService: UserService) {
     @DeleteMapping("/user/{id}")
     @Operation(summary = "Удаление пользователя")
     fun delete(@CookieValue("jwt") jwt: String?, @PathVariable id: Int): ResponseEntity<Any> {
+
+        if (!ftService.isEnabled(userCanDelete)) {
+            println("feature $userCanDelete is off")
+            return ResponseEntity.ok().body(Message("FT $userCanDelete is off"))
+        }
+        println("feature $userCanDelete is on")
+
         return userService.deleteUser(id)
     }
 
     @GetMapping("/user")
     @Operation(summary = "Получение данных о пользователе")
     fun getUser(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+
+        if (!ftService.isEnabled(userCanGet)) {
+            println("feature $userCanGet is off")
+            return ResponseEntity.ok().body(Message("FT $userCanGet is off"))
+        }
+        println("feature $userCanGet is on")
+
         try {
             if (jwt == null) {
                 return ResponseEntity.status(401).body(Message("unauthorized"))
             }
 
+            println("feature $userCanGet is on")
             val body = Jwts.parser()
                 .setSigningKey("secretsecretsecretsecretsecretsecretsecretsecret") // TODO Вынести в конфигурацию
                 .parseClaimsJws(jwt).body
@@ -81,6 +119,13 @@ class AuthController(private val userService: UserService) {
     @PostMapping("/login")
     @Operation(summary = "Авторизация пользователя")
     fun login(@RequestBody body: UserDTO, response: HttpServletResponse): ResponseEntity<Any> {
+
+        if (!ftService.isEnabled(userCanLogin)) {
+            println("feature $userCanLogin is off")
+            return ResponseEntity.ok().body(Message("FT $userCanLogin is off"))
+        }
+        println("feature $userCanLogin is on")
+
         val user = this.userService.findByEmail(body.email)
             ?: return ResponseEntity.badRequest().body(Message("user not found"))
 
@@ -117,11 +162,16 @@ class AuthController(private val userService: UserService) {
     @PostMapping("/logout")
     @Operation(summary = "Деавторизация пользователя")
     fun logout(response: HttpServletResponse): ResponseEntity<Any> {
+
+        if (!ftService.isEnabled(userCanlogout)) {
+            println("feature $userCanlogout is off")
+            return ResponseEntity.ok().body(Message("FT $userCanlogout is off"))
+        }
+        println("feature $userCanLogin is on")
+
         val cookie = Cookie("jwt", "")
         cookie.maxAge = 0
-
         response.addCookie(cookie)
-
         return ResponseEntity.ok(Message("success"))
     }
 }
